(subtree) Modify the pre-receive hook to deal with merges
Previously the pre-receive hook would reject merge commits because
technically they cause a diff that shows multiple module files changing.
This commit makes the hook more or less ignore merge commits, trusting
that the merger was sane. Dumb, but we need to be able to push merge
commits so making the change for now.