(subtree) Fix merge-base check in pre-receive hook
Previously the merge-base check in the pre-receive hook used the oldrev
as the rev to check the merge base against. This doesn't actually do
what it's supposed to do.

The idea behind the merge-base check is that any commit which doesn't
share a merge base with the root commit of the master branch shouldn't
be checked by this hook. This allows for git-subtree pulls to function
without having to manually bypass the pre-receive hook.

This commit adds an extra parameter to the pre-receive hook,
MASTER_BASE, which should be set to the commit hash of the base commit
of the master branch in the repository. All merge-base checks will be
performed against this commit, and so will function as intended.